+++ src/main/java/org/http/HttpRetryAware.java
@@ -1,0 +1,21 @@
+package org.http;
+
+/**
+ * http重试
+ * @author zouziwen
+ *
+ */
+public interface HttpRetryAware {
+	
+	/**
+	 * 是否重试
+	 * @return
+	 */
+	boolean isRetry();
+	
+	/**
+	 * 重试的具体实施策略
+	 * @param handler
+	 */
+	HttpRequestRetryHandler retryHandler();
+}