+++ CHANGELOG.md
@@ -32,7 +32,6 @@ CHANGELOG
 
 1. 添加Releaseable接口，用于释放request资源 
 2. 添加请求参数时当值为空的时候也可以添加
+3. 重试策略以接口形式HttpRetryAware
+4. 将request和response的关闭开放
 
-
-
-